refactor(gms): separate shared core from v1 policy
