+++ lib/gpu_memory_service/integrations/sglang/memory_saver.py
@@ -26,8 +26,8 @@ from gpu_memory_service.client.torch.allocator import (
     get_or_create_gms_client_memory_manager,
     gms_use_mem_pool,
 )
-from gpu_memory_service.common.locks import GrantedLockType, RequestedLockType
 from gpu_memory_service.common.utils import GMS_TAGS, get_socket_path
+from gpu_memory_service.core.locks import GrantedLockType, RequestedLockType
 from gpu_memory_service.integrations.common.utils import finalize_gms_write
 
 logger = logging.getLogger(__name__)
