+++ lib/gpu_memory_service/integrations/vllm/model_loader.py
@@ -23,8 +23,8 @@ from gpu_memory_service.client.torch.module import (
     materialize_module_from_gms,
     rebind_nonparameter_tensors,
 )
-from gpu_memory_service.common.locks import GrantedLockType
 from gpu_memory_service.common.utils import get_socket_path
+from gpu_memory_service.core.locks import GrantedLockType
 from gpu_memory_service.integrations.common.utils import (
     GMSCommittedMemoryStats,
     get_gms_lock_mode,
