+++ lib/gpu_memory_service/v1/cli.py
@@ -13,8 +13,7 @@ from threading import Event
 from gpu_memory_service.common.utils import get_socket_path
 from gpu_memory_service.common.vmm import get_vmm
 from gpu_memory_service.core import device as device_identity
-from gpu_memory_service.core.server.gms import GMSServerMemoryManager
-from gpu_memory_service.core.server.rpc import GMSRPCServer
+from gpu_memory_service.v1.server import GMSRPCServer, GMSServerMemoryManager
 
 _DOMAINS = ("weights", "kv_cache")
